fastapi/requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.9.1
fastapi==0.61.2
orjson==3.7.7
pydantic==1.9.0
uvicorn==0.12.2
uvloop==0.16.0
flake8==4.0.1
gunicorn==20.1.0
redis==4.3.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.28.1
